Final_Project_PROG2200_Cullen_MB/alert-notification-system/src/test/java/com/example/alertnotificationsystem/controller/PersonControllerTest.java
package com.example.alertnotificationsystem.controller;

import com.example.alertnotificationsystem.service.PersonService;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.web.servlet.MockMvc;

import java.util.List;
import java.util.Map;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@WebMvcTest(PersonController.class)
public class PersonControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private PersonService personService; // Mock the dependency

    @Test
    public void testGetAllPersons() throws Exception {
        when(personService.getAllPersons()).thenReturn(
                List.of(
                        Map.of("id", 1L, "firstName", "John", "lastName", "Doe", "address", "123 Main St")
                )
        );

        mockMvc.perform(get("/api/persons"))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$[0].firstName").value("John"))
                .andExpect(jsonPath("$[0].lastName").value("Doe"))
                .andExpect(jsonPath("$[0].address").value("123 Main St"));
    }
}
